+++ cmd/clawdex/main.go
@@ -8,8 +8,10 @@ import (
 	"github.com/openclaw/clawdex/internal/cli"
 )
 
+var exit = os.Exit
+
 func main() {
-	os.Exit(run(os.Args[1:], os.Stdout, os.Stderr))
+	exit(run(os.Args[1:], os.Stdout, os.Stderr))
 }
 
 func run(args []string, stdout, stderr io.Writer) int {